refactor: separate general logic into utils
