+++ libs/vscode/nx-project-view/src/lib/views/index.ts
@@ -1,0 +1,2 @@
+export * from './nx-project-list-view';
+export * from './nx-project-tree-view';
